EES-4048 Refactor out `SelectedRelease` type for use in data set step
See EES-4312 for ticket to refactor `SelectedPublication` and `SelectedRelease`
further.
