+++ src/explore-education-statistics-frontend/src/modules/table-tool/__tests__/TableToolPage.test.tsx
@@ -1,7 +1,7 @@
+import { SelectedPublication } from '@common/modules/table-tool/types/selectedPublication';
 import { Theme } from '@common/services/publicationService';
 import {
   FastTrackTable,
-  SelectedPublication,
   Subject,
   SubjectMeta,
 } from '@common/services/tableBuilderService';
